Rename : req-memo 라우트 components 폴더 추가
- ClientCatComponent, ServerCatComponent 이동
